requirements.txt
discord.py>=2.3.0
supabase>=2.10.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
cachetools>=5.3.0
pydantic>=2.0.0
